update lib gen script

+++ platform/tools/genlib/genlib.sh
@@ -22,8 +22,15 @@ MAKE_J_NUMBER=`cat /proc/cpuinfo | grep vendor_id | wc -l`
 echo "==core number:$MAKE_J_NUMBER=="
 make -j${MAKE_J_NUMBER} CT_RELEASE=$compileMode 2>&1
 mkdir -p hex/libcsdk
-cp -f build/csdk/lib/libcsdk_${compileMode}.a hex/libcsdk/libcsdk_${compileMode}.a
-echo "Generate lib OK"
+if [[ -f "build/csdk/lib/libcsdk_${compileMode}.a" ]]; then
+    cp -f build/csdk/lib/libcsdk_${compileMode}.a hex/libcsdk/libcsdk_${compileMode}.a
+    echo "Generate lib OK"
+else
+    echo "Generate lib FAIL"
+fi
+
+
+
 
 ###############################
 mv -f Makefile.bak Makefile
